Move all code for compute_reference_metrics out of BasicMetric.
My hope is to simplify BasicMetric to the point where it can be broken up into more comprehensible pieces.
